Implement RIPE Atlas Stream API
- Add Stream API Atlaser
- Moved example to examples, add streamer example
