+++ .travis.yml
@@ -3,3 +3,4 @@ go:
   - master
 install:
   - go get github.com/miekg/dns
+  - go get github.com/graarh/golang-socketio
